pyproject.toml
[project]
name = "professor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.49.1",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langsmith>=0.4.41",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
